refactor(refactor weather_data): refactor in modo da rispettare architettura a microservizi
